+++ filchain/filchain.go
@@ -1,19 +1,16 @@
-package lotuschain
+package filchain
 
 import (
 	"context"
 	"fmt"
 
 	"github.com/filecoin-project/lotus/chain/types"
-	"github.com/textileio/powergate/ffs/filcold"
 )
 
-// LotusChain is an implementation of FilChain which targets a Lotus node.
-type LotusChain struct {
+// FIlChain is an abstraction of the Filecoin network.
+type FilChain struct {
 	api API
 }
-
-var _ filcold.FilChain = (*LotusChain)(nil)
 
 // API interacts with a Lotus full-node
 type API interface {
@@ -21,14 +18,14 @@ type API interface {
 }
 
 // New creates a new deal module
-func New(api API) *LotusChain {
-	return &LotusChain{
+func New(api API) *FilChain {
+	return &FilChain{
 		api: api,
 	}
 }
 
 // GetHeight returns the current height of the chain for the targeted Lotus node.
-func (lc *LotusChain) GetHeight(ctx context.Context) (uint64, error) {
+func (lc *FilChain) GetHeight(ctx context.Context) (uint64, error) {
 	h, err := lc.api.ChainHead(ctx)
 	if err != nil {
 		return 0, fmt.Errorf("get head from lotus node: %s", err)
